refactor(bunny): extract pure URL helpers into bunny-urls.ts
URL transformers (getOptimizedImageUrl, getThumbnailUrl, …) have no
Convex runtime dependency and are candidates for frontend reuse. Move
them into a dedicated module; bunny.ts re-exports them so existing
imports keep working.
